Changed strategy of how to garner & store the data
changed the returns_table test, still not complete.
changed the cook_recipe() -- still uncomplete -- into getting the data
from the two rows that are containing it.
add that data to a list as well as created a list with keys that will be
merged with the aforementioned list to create a dict.