Building the Freezer class
Freezer class will connect to the mysql db and be a storage for the bot.
Freezer.__init__ initializes a peewee.MySQLDatabase object using the provided args.
Freezer.open_freezer(self) connects to the database.
Freezer.close_freezer(self) closes the connection to the database.

test/test_freezer contains the tests for the Freezer class, currently
only the opening & closing of a database connection is being tested.

all tests pass