Changing strategy to capturing the entire html
Since the content providers changed their website after this was built,
I have decided to chage the strategy in which I provide the information
from them. From hence forth, everything will be removed and instead only
user data will be saved in the db. I will present the entire html to the
users on telegram and build the bot accordingly.

* re-format how exceptions are placed in cooker.get_recipe()
* add get_spice_results() that will get the results of yesterday's
matches
* add update_bot_content() to freezer
* add winlose to PlatiniumBotContent
* add new file test_code_idea to try out ideas on
* add test_add_spice_results() to test_cooker
* import timezone in test_freezer
* correct test_add_new_content() in test_freezer
* add a few more assertions to test_add_new_content() in test_freezer
* add test_get_bot_content_yesterday() in test_freezer
* add add_content_yesterday_for_testing() in test_freezer